Enable metrics on Deer app servers
- Also add JVM metrics to both app and processing servers
